added correct range of EBSAir sensors, added docs too

+++ Core/Inc/ADCTask.h
@@ -19,9 +19,9 @@ typedef struct{
     uint16_t desmo2;
     uint16_t clutchOil;
     uint16_t GearUpAir;
-    uint16_t APPS1;
-    uint16_t APPS2;
-    uint16_t APPS3;
+    uint8_t APPS1;
+    uint8_t APPS2;
+    uint8_t APPS3;
     uint16_t VPPMSense;
     uint16_t BPSFront;
     uint16_t BPSRear;
